requirements.txt
bokeh==2.2.3
click==7.1.2
DateTime==4.3
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.0
packaging==20.9
pandas==1.2.1
Pillow==8.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.4.1
six==1.15.0
tornado==6.1
typing-extensions==3.7.4.3
Werkzeug==1.0.1
zope.interface==5.2.0
